put kernel source in differents files + clean the test into one test fil
