+++ pytissueoptics/rayscattering/opencl/example.py
@@ -18,7 +18,7 @@ These parameters will be used to mimic the parameters a typical user would utili
 
 
 tissue = CubeTissue(side=10, material=ScatteringMaterial(30, 0.1, 0.8, 1.4))
-source = PencilSource(position=Vector(0, 0, 0), direction=Vector(0, 0, 1), N=20000, use_opencl=True)
+source = PencilSource(position=Vector(0, 0, 0), direction=Vector(0, 0, 1), N=50000, use_opencl=True)
 logger = Logger()
 
 source.propagate(tissue, logger=logger)
